fix ci: remove globalEntityRegistryV2 cyclic import; remove RelatedEntity (unused); clean up AccessButtonHelpers.test.tsx; fix imports
