+++ datahub-web-react/src/app/entityV2/shared/tabs/Dataset/AccessManagement/AccessButtonHelpers.tsx
@@ -7,13 +7,13 @@ import { ANTD_GRAY, REDESIGN_COLORS } from '@app/entityV2/shared/constants';
 /**
  * Tooltip message for when user already has access
  */
-const ACCESS_GRANTED_TOOLTIP = 'You already have access to this role';
+export const ACCESS_GRANTED_TOOLTIP = 'You already have access to this role';
 
 /**
  * Styled button component for access management actions.
  * Supports both enabled (request) and disabled (granted) states.
  */
-const AccessButton = styled(Button)`
+export const AccessButton = styled(Button)`
     background-color: ${REDESIGN_COLORS.BLUE};
     color: ${REDESIGN_COLORS.WHITE};
     width: 80px;
@@ -53,10 +53,20 @@ export interface RoleAccessData {
 }
 
 /**
+ * Returns the button text based on whether the user has access.
+ */
+export const getAccessButtonText = (hasAccess: boolean): string => (hasAccess ? 'Granted' : 'Request');
+
+/**
+ * Returns whether the access button should be disabled.
+ */
+export const isAccessButtonDisabled = (hasAccess: boolean): boolean => hasAccess;
+
+/**
  * Handles the click event for access request buttons.
  * Only opens the URL if the user doesn't already have access.
  */
-const handleAccessButtonClick = (hasAccess: boolean, url?: string) => (e: React.MouseEvent) => {
+export const handleAccessButtonClick = (hasAccess: boolean, url?: string) => (e: React.MouseEvent) => {
     if (!hasAccess && url) {
         e.preventDefault();
         window.open(url);
