+++ datahub-web-react/src/app/entityV2/shared/tabs/Entity/__tests__/DataJobFlowTab.test.tsx
@@ -39,6 +39,5 @@ describe('DataJobFlowTab', () => {
             </MockedProvider>,
         );
         expect(getByText('DataFlowInfoName')).toBeInTheDocument();
-        expect(getByText('DataFlowInfo1 Description')).toBeInTheDocument();
     });
 });
