+++ datahub-web-react/src/app/lineageV2/useComputeGraph/orderNodes.ts
@@ -1,4 +1,4 @@
-import { globalEntityRegistryV2 } from '@app/EntityRegistryProvider';
+import globalEntityRegistryV2 from '@app/globalEntityRegistryV2';
 import { LineageEntity, NodeContext, isUrnTransformational } from '@app/lineageV2/common';
 
 import { LineageDirection } from '@types';
