--- datahub-web-react/src/app/useBuildEntityRegistry.ts
@@ -1,9 +1,0 @@
-import { useMemo } from 'react';
-
-import buildEntityRegistryV2 from '@app/buildEntityRegistryV2';
-
-export default function useBuildEntityRegistry() {
-    return useMemo(() => {
-        return buildEntityRegistryV2();
-    }, []);
-}
